EduEase/requirements.txt
streamlit
faster-whisper
openai
ctransformers
huggingface-hub
gTTS
Pillow
yt-dlp
ffmpeg-python